requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via litellm
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   openai
attrs==25.4.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
cachetools==6.2.4
    # via zai-sdk
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   litellm
    #   typer-slim
distro==1.9.0
    # via openai
fastuuid==0.14.0
    # via litellm
filelock==3.20.2
    # via huggingface-hub
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.12.0
    # via huggingface-hub
grpcio==1.67.1
    # via litellm
h11==0.16.0
    # via httpcore
hf-xet==1.2.0
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   huggingface-hub
    #   litellm
    #   openai
    #   zai-sdk
huggingface-hub==1.2.3
    # via tokenizers
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.1
    # via litellm
jinja2==3.1.6
    # via litellm
jiter==0.12.0
    # via openai
jsonschema==4.25.1
    # via litellm
jsonschema-specifications==2025.9.1
    # via jsonschema
litellm==1.80.11
    # via geo (pyproject.toml)
markupsafe==3.0.3
    # via jinja2
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
openai==2.14.0
    # via litellm
packaging==25.0
    # via huggingface-hub
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pydantic==2.12.5
    # via
    #   geo (pyproject.toml)
    #   litellm
    #   openai
    #   zai-sdk
pydantic-core==2.41.5
    # via
    #   pydantic
    #   zai-sdk
pyjwt==2.10.1
    # via zai-sdk
python-dotenv==1.2.1
    # via litellm
pyyaml==6.0.3
    # via huggingface-hub
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2025.11.3
    # via tiktoken
requests==2.32.5
    # via tiktoken
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
shellingham==1.5.4
    # via huggingface-hub
sniffio==1.3.1
    # via openai
tiktoken==0.12.0
    # via litellm
tokenizers==0.22.2
    # via litellm
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
typer-slim==0.21.0
    # via huggingface-hub
typing-extensions==4.15.0
    # via
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   typer-slim
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.2
    # via requests
yarl==1.22.0
    # via aiohttp
zai-sdk==0.2.0
    # via geo (pyproject.toml)
zipp==3.23.0
    # via importlib-metadata
